# openapi/sovereign.v1.yaml
openapi: 3.1.0
info:
  title: Sovereign REST API
  version: "1.0.0"
  description: |
    Control-plane API for Sovereign tenants. Hand-written source of truth
    (D-5): the route set in internal/api/router.go must stay in parity with
    this spec in both directions, enforced by the OpenAPI drift test.
servers:
  - url: https://id.<domain>/api/v1
    description: Identity host, versioned API base.
# Default security for the API: bearer (programmatic) or session cookie
# (browser). Operations that are anonymous override this with `security: []`.
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /me/profile:
    get:
      operationId: getSelfProfile
      summary: Get the current principal's profile page
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Profile page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfilePage"
        "404":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
    put:
      operationId: upsertSelfProfile
      summary: Create or fully replace the profile page
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpsertRequest"
      responses:
        "200":
          description: Upserted profile page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfilePage"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteSelfProfile
      summary: Delete the profile page (cascades to its links)
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "204":
          description: Profile deleted
        "401":
          $ref: "#/components/responses/Problem"
  /me/profile:publish:
    post:
      operationId: publishSelfProfile
      summary: Publish the profile page
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Published profile page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfilePage"
        "404":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
  /me/profile:unpublish:
    post:
      operationId: unpublishSelfProfile
      summary: Unpublish the profile page
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Unpublished profile page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfilePage"
        "404":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: array
          maxItems: 10
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      required: [field, code]
      properties:
        field:
          type: string
        code:
          type: string
        detail:
          type: string
...
    ProfilePage:
      type: object
      properties:
        id:
          type: string
        display_name:
          type: string
        bio:
          type: string
        is_published:
          type: boolean
        avatar_url:
          type: string
          format: uri
        theme:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProfileUpsertRequest:
      type: object
      properties:
        display_name:
          type: string
          maxLength: 100
        bio:
          type: string
        is_published:
          type: boolean
...